Time: 1190 ms (7.68%), Space: 423.2 MB (7.51%) - LeetHub

+++ 2170-count-number-of-maximum-bitwise-or-subsets/2170-count-number-of-maximum-bitwise-or-subsets.cpp
@@ -32,7 +32,7 @@
 
 class Solution {
 public:
-    int solve (int idx,int curOr,vector<int> nums,int sizeOfArray,int maxOr)
+    int solve (int idx,int curOr,vector<int> nums,short sizeOfArray,int maxOr)
     {
         if(idx==sizeOfArray)
         return maxOr==curOr;
@@ -49,7 +49,7 @@ public:
         {
             maxOr |= val;
         }
-        int n=nums.size();
+        short n=nums.size();
         return solve(0,0,nums,n,maxOr); 
     }
 };
